--- src/myteam/constants.py
@@ -1,11 +1,0 @@
-APP_NAME = "myteam"
-DEFAULT_ROLE = "main"
-AGENTS_DIRNAME = ".myteam"
-ENCODING = "utf-8"
-INSTRUCTIONS_FILE = "instructions.md"
-INFO_FILE = "info.md"
-AGENT_PY_FILE = "agent.py"
-AGENTS_MD_FILE = "AGENTS.md"
-ROSTER_REPOSITORY_URL = "https://api.github.com/repos/beanlab/rosters/git/trees"
-ROSTER_RAW_BASE_URL = "https://raw.githubusercontent.com/beanlab/rosters/refs/heads/main"
-ZIP_FILE_NAME = "roster.zip"
